chore: Comment out unnecessary code for training
The code changes in this commit involve commenting out the lines that run the `train.py` script and empty the CUDA cache in the `train.ipynb` file. These changes are made to disable the training process and prevent the CUDA cache from being cleared during notebook execution. This is done to avoid unnecessary execution and improve the notebook's performance.